Convert some stragglers to size_t
I wasn't really planning on doing this, but everything else is size_t
now, so I might as well move these two over as well to elminate some
'comparing variabels of differing sign' warnings.

+++ SonicCDDecomp/Audio.hpp
@@ -30,12 +30,12 @@ struct MusicPlaybackInfo {
 struct SFXInfo {
     char name[0x40];
     Sint16 *buffer;
-    int length;
+    size_t length;
     bool loaded;
 };
 
 struct ChannelInfo {
-    int sampleLength;
+    size_t sampleLength;
     Sint16 *samplePtr;
     int sfxID;
     byte loopSFX;
